This is another quick-ish fix to less cbor encode/decode.

Previously the "Value" of a KV was stored as a cbor encoded interface{} which meant that everytime you do a write to the hamt, every node that is touched is deserializing those *values* into the correct objects, even though your app has no use for those objects. This changes the Value to be a []byte which won't get serialized/deserialized except when its getting stored or when its found. This has way more of an impact on Community:

Before:
(community Benchmark)
```
Running tool: /usr/local/bin/go test -benchmem -run=^$ github.com/quorumcontrol/community/hub/tupelocache -bench ^(BenchmarkAddAddBlockRequest)$

goos: darwin
goarch: amd64
pkg: github.com/quorumcontrol/community/hub/tupelocache
BenchmarkAddAddBlockRequest-12    	    2000	   3333451 ns/op	 1457153 B/op	   24186 allocs/op
PASS
ok  	github.com/quorumcontrol/community/hub/tupelocache	8.284s
Success: Benchmarks passed.
```

After:
```
Running tool: /usr/local/bin/go test -benchmem -run=^$ github.com/quorumcontrol/community/hub/tupelocache -bench ^(BenchmarkAddAddBlockRequest)$

goos: darwin
goarch: amd64
pkg: github.com/quorumcontrol/community/hub/tupelocache
BenchmarkAddAddBlockRequest-12    	    3000	   1696724 ns/op	  883177 B/op	   11961 allocs/op
PASS
ok  	github.com/quorumcontrol/community/hub/tupelocache	7.376s
Success: Benchmarks passed.
```